requirements.txt
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Jinja2==2.10.1
numpy==1.16.3
opencv-python==4.1.0.25